requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs-19.1.0
chardet==3.0.2
discord.py==1.2.2
idna-ssl==1.1.0
fuzzywuzzy==0.15.0
multidict==4.5.2
typing-extensions==3.7.2
numexpr==2.6.2
numpy==1.12.1
python-Levenshtein==0.12.0
virtualenv==15.1.0
websockets==6.0
yarl==1.3.0
redis==2.10.6
pytz==2017.3
pygsheets
oauth2client
markovify==0.7.0
Pillow
mcstatus
requests
websocket-client
